StringDiskDeserializaer using UTF16
