--- Explorer/Assets/DCL/Infrastructure/SceneRuntime/Factory/JsSource/StringDiskSerializer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4a83e736de9d4878b6e5d9923bd3449a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
